minor tiny changes to the etl.py script
This is the initial commit of the explore_pymc.py
-- contains some sample code to simulate histrograms
-- contains a dtereminitic skeleton to solve an SEIR-c model which will be used to generate data further in this analysis
-- contains a wrapper function that integrates the system.

Note:
This script is going to contain a disease model which I intend to conduct simulation based parameter estimation on using pymc.
The objective is to provide pymc with a sufficently diffciult problem with the known solution to identify bottlenecks in analyses that contain the some complex structures like diff equations for example.